fix export state
